# Conflicts:
#	build.gradle
#	src/main/java/com/devloop/common/apipayload/status/ErrorStatus.java
#	src/main/java/com/devloop/communitycomment/service/CommunityCommentService.java
#	src/main/java/com/devloop/config/WebSecurityConfig.java
#	src/main/java/com/devloop/partycomment/service/PartyCommentService.java
#	src/main/resources/application.yaml

+++ src/main/java/com/devloop/community/service/CommunityService.java
@@ -1,34 +1,38 @@
 package com.devloop.community.service;
 
+import com.devloop.attachment.entity.CommunityAttachment;
+import com.devloop.attachment.repository.CommunityATMRepository;
+import com.devloop.attachment.s3.S3Service;
+import com.devloop.attachment.service.CommunityAttachmentService;
 import com.devloop.common.AuthUser;
+import com.devloop.common.apipayload.dto.CommunitySimpleResponseDto;
 import com.devloop.common.apipayload.status.ErrorStatus;
 import com.devloop.common.enums.BoardType;
 import com.devloop.common.enums.Category;
 import com.devloop.common.exception.ApiException;
 import com.devloop.common.utils.SearchResponseUtil;
-import com.devloop.community.dto.request.CommunitySaveRequest;
-import com.devloop.community.dto.request.CommunityUpdateRequest;
-import com.devloop.community.dto.response.CommunityDetailResponse;
-import com.devloop.community.dto.response.CommunitySaveResponse;
-import com.devloop.community.dto.response.CommunitySimpleResponse;
+import com.devloop.community.request.CommunitySaveRequest;
+import com.devloop.community.request.CommunityUpdateRequest;
+import com.devloop.community.response.CommunityDetailResponse;
+import com.devloop.community.response.CommunitySaveResponse;
+import com.devloop.community.response.CommunitySimpleResponse;
 import com.devloop.community.entity.Community;
 import com.devloop.community.entity.ResolveStatus;
 import com.devloop.community.repository.CommunityRepository;
-import com.devloop.communitycomment.repository.CommunityCommentRepository;
 import com.devloop.search.response.IntegrationSearchResponse;
 import com.devloop.user.entity.User;
-import com.devloop.user.repository.UserRepository;
+import com.devloop.user.service.UserService;
 import lombok.RequiredArgsConstructor;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.PageRequest;
 import org.springframework.data.domain.PageImpl;
-import org.springframework.data.domain.PageRequest;
 import org.springframework.data.domain.Pageable;
 import org.springframework.data.jpa.domain.Specification;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
+import org.springframework.web.multipart.MultipartFile;
 
-import java.util.ArrayList;
+import java.net.URL;
 import java.util.List;
 
 @Service
@@ -36,15 +40,17 @@ import java.util.List;
 @Transactional(readOnly = true)
 public class CommunityService {
     private final CommunityRepository communityRepository;
-    private final UserRepository userRepository; //서비스로 받아오게
+    private final UserService userService;
+    private final S3Service s3Service;
+    private final CommunityATMRepository communityATMRepository;
+    private final CommunityAttachmentService communityAttachmentService;
 
     //게시글 작성
     @Transactional
-    public CommunitySaveResponse createCommunity(AuthUser authUser, CommunitySaveRequest communitySaveRequest) {
+    public CommunitySaveResponse createCommunity(AuthUser authUser, MultipartFile file, CommunitySaveRequest communitySaveRequest) {
         Category category = Category.of(communitySaveRequest.getCategory());
         //사용자 조회
-        User user = userRepository.findById(authUser.getId())
-                .orElseThrow(() -> new ApiException(ErrorStatus._NOT_FOUND_USER));
+        User user = userService.findByUserId(authUser.getId());
         //게시글 Community객체 생성
         Community community = Community.of(
                 communitySaveRequest.getTitle(),
@@ -53,6 +59,10 @@ public class CommunityService {
                 user);
         //게시글 저장
         Community savedCommunity = communityRepository.save(community);
+        //첨부파일 있으면 저장
+        if (file!=null && !file.isEmpty()){
+            s3Service.uploadFile(file,user,community); //s3에 파일 올리고 communityattachment에 저장하는 것
+        }
         //응답반환
         return CommunitySaveResponse.of(
                 savedCommunity.getId(),
@@ -67,8 +77,16 @@ public class CommunityService {
     //게시글 다건 조회
     public Page<CommunitySimpleResponse> getCommunities(int page, int size) {
         Pageable pageable = PageRequest.of(page - 1, size);
-        //페이지네이션된 게시글 조회
-        return communityRepository.findAllSimple(pageable);
+        //페이지네이션된 게시글 조회하고 응답
+        Page<CommunitySimpleResponseDto> communityDtos = communityRepository.findAllSimple(pageable)
+                .orElseThrow(() -> new ApiException(ErrorStatus._NOT_FOUND_COMMUNITY));
+
+        return communityDtos.map(dto -> CommunitySimpleResponse.of(
+                dto.getCommunityId(),
+                dto.getTitle(),
+                dto.getStatus().getDescription(),
+                dto.getCategory().getDescription()
+        ));
     }
 
     //게시글 단건(상세조회)
@@ -76,27 +94,38 @@ public class CommunityService {
         //게시글 조회
         Community community = communityRepository.findById(communityId)
                 .orElseThrow(() -> new ApiException(ErrorStatus._NOT_FOUND_COMMUNITY));
-
+        //첨부파일 url이 있는지 확인
+        String imageUrl = communityATMRepository.findByCommunityId(communityId) //첨부파일 있는지 조회
+                .map(CommunityAttachment::getImageURL)
+                .map(URL ::toString)
+                .orElse(null);
         //응답반환
-        return CommunityDetailResponse.of(
+        return CommunityDetailResponse.withAttachment(
                 community.getId(),
                 community.getTitle(),
                 community.getContent(),
                 community.getResolveStatus().getDescription(),
                 community.getCategory().getDescription(),
                 community.getCreatedAt(),
-                community.getModifiedAt()
+                community.getModifiedAt(),
+                imageUrl
         );
     }
 
     //게시글 수정
     @Transactional
-    public CommunityDetailResponse updateCommunity(Long communityId, CommunityUpdateRequest communityUpdateRequest) {
+    public CommunityDetailResponse updateCommunity(AuthUser authUser,Long communityId, CommunityUpdateRequest communityUpdateRequest, MultipartFile file) {
         ResolveStatus resolvedStatus = ResolveStatus.of(communityUpdateRequest.getStatus());
         Category category = Category.of(communityUpdateRequest.getCategory());
+
         //게시글 조회
         Community community = communityRepository.findById(communityId)
                 .orElseThrow(() -> new ApiException(ErrorStatus._NOT_FOUND_COMMUNITY));
+
+        //작성자 확인
+        if (!community.getUser().getId().equals(authUser.getId())) {
+            throw new ApiException(ErrorStatus._PERMISSION_DENIED);
+        }
 
         //수정 요청에서 값이 있는 필드만 업데이트시키기
         community.updateCommunity(
@@ -105,10 +134,19 @@ public class CommunityService {
                 resolvedStatus,
                 category
         );
-        //수정된 게시글 저장
-        communityRepository.save(community);
+
+        //첨부파일 수정
+        if (file != null && !file.isEmpty()) {
+            // 기존 파일이 있는지 확인
+            CommunityAttachment communityAttachment = communityAttachmentService.findCommunityAttachmentByCommunityId(communityId);
+            if (communityAttachment == null) {
+                s3Service.uploadFile(file, community.getUser(), community);
+            } else {
+                s3Service.updateUploadFile(file, communityAttachment, community);
+            }
+        }
         //응답반환
-        return CommunityDetailResponse.of(
+        return CommunityDetailResponse.withoutAttachment(
                 community.getId(),
                 community.getTitle(),
                 community.getContent(),
@@ -121,10 +159,19 @@ public class CommunityService {
 
     //게시글 삭제
     @Transactional
-    public void deleteCommunity(Long communityId) {
+    public void deleteCommunity(AuthUser authUser,Long communityId) {
         //게시글 존재하는지 확인
         Community community = communityRepository.findById(communityId)
                 .orElseThrow(() -> new ApiException(ErrorStatus._NOT_FOUND_COMMUNITY));
+        //작성자 확인
+        if (!community.getUser().getId().equals(authUser.getId())) {
+            throw new ApiException(ErrorStatus._PERMISSION_DENIED);
+        }
+        // 첨부파일 확인 및 삭제
+        communityATMRepository.findByCommunityId(communityId).ifPresent(attachment -> {
+            s3Service.delete(attachment.getFileName());
+            communityATMRepository.delete(attachment);
+        });
         //삭제
         communityRepository.delete(community);
     }
